cast unlocking
very tenuous. Making a new spell requires creating it in the sky circle, implementing the call in the main level code, creating the call functionality wherever, and setting up unlocks in the star chart by creating a resource and adding it to the dictionary and unlock arrays. OOF. Also, for unlocks you gotta update the main level unlock list. This shit is a nightmare, but I'm not spending too much longer on this project so I don't really care.